4. Тесты для стороны прямоугольника. [#429280]

+++ src/ru/job4j/condition/SqArea.java
@@ -1,7 +1,7 @@
 package ru.job4j.condition;
 
 public class SqArea {
-    public static double square(int p, int k) {
+    public static double square(double p, int k) {
         return Math.pow((p / (2 * (k + 1))), 2) * k;
     }
